+++ src/app/(frontend)/(home)/components/InformationPanel.tsx
@@ -1,5 +1,6 @@
 // import { satoshi } from "../../layout";
 import { tv, VariantProps } from "tailwind-variants";
+
 import { cn } from "@/libs/utils";
 
 const informationPanel = tv({
@@ -7,13 +8,13 @@ const informationPanel = tv({
     variants: {
         headerColour: {
             purple: "text-purple-border",
-            pink: "text-pink-border"
+            pink: "text-pink-border",
         },
     },
     defaultVariants: {
         headerColour: "purple",
-    }
-})
+    },
+});
 
 type InformationPanelProps = {
     title: String;
@@ -22,9 +23,9 @@ type InformationPanelProps = {
     variant?: VariantProps<typeof informationPanel>;
 };
 
-export default function InformationPanel({ title, subtitle, description, variant }: InformationPanelProps) {
+export function InformationPanel({ title, subtitle, description, variant }: InformationPanelProps) {
     return (
-        <div className="w-106 flex flex-col gap-4">
+        <div className="flex w-106 flex-col gap-4">
             <h1 className={cn(informationPanel(variant))}>{title}</h1>
             <h3 className="font-tomorrow text-3xl font-normal">{subtitle}</h3>
             <p className="font-satoshi font-light">{description}</p>
