+++ src/app/(frontend)/globals.css
@@ -10,12 +10,12 @@
 @theme {
     --color-purple: #7650ff;
     /* --color-purple-border: #5d3dd1; */
-    --color-purple-border: #4B63FF;
-    --color-purple-button: #4B63FF99;
+    --color-purple-border: #4b63ff;
+    --color-purple-button: #4b63ff99;
     --color-gray: #1e1e1e;
     --color-gray-border: #363636;
-    --color-pink: #FF4B9966;
-    --color-pink-border: #FF4B99;
+    --color-pink: #ff4b9966;
+    --color-pink-border: #ff4b99;
 
     --font-tomorrow: "tomorrow", sans-serif;
     --font-satoshi: "satoshi", sans-serif;
@@ -62,12 +62,8 @@ h3 {
     line-height: 91.5%;
     font-weight: 500;
 }
-
-p {
-    letter-spacing: 4%;
-    /* line-height: 130%; */
-    /* font-weight: 500; */
-}
+    
+*/
 
 p.font-bold,
 p em {
@@ -85,8 +81,8 @@ small {
 small.font-bold,
 small em {
     letter-spacing: -2%;
-    font-weight: 500; 
-}*/
+    font-weight: 500;
+}
 
 @media screen and (min-width: 1024px) {
     p {
