--- src/app/(frontend)/landing/_component/icons/TwitchIcon.tsx
@@ -1,28 +1,0 @@
-import React from "react";
-
-const TwitchIcon = () => {
-    return (
-        <svg
-            width="32"
-            height="32"
-            viewBox="0 0 32 32"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-        >
-            <rect width="32" height="32" fill="white" />
-            <g clip-path="url(#clip0_1017_312)">
-                <path
-                    d="M22.9 5.125H26.581L18.541 14.338L28 26.875H20.5945L14.794 19.27L8.1565 26.875H4.474L13.0735 17.02L4 5.125H11.5945L16.837 12.0745L22.9 5.125ZM21.61 24.667H23.65L10.4845 7.2175H8.2975L21.61 24.667Z"
-                    fill="black"
-                />
-            </g>
-            <defs>
-                <clipPath id="clip0_1017_312">
-                    <rect width="24" height="24" fill="white" transform="translate(4 4)" />
-                </clipPath>
-            </defs>
-        </svg>
-    );
-};
-
-export { TwitchIcon };
