+++ tsconfig.json
@@ -19,7 +19,6 @@
             }
         ],
         "paths": {
-            "@ui/*": ["./src/components/ui/*"],
             "@schema": ["./src/libs/db/schema/index.ts"],
             "@providers/*": ["./src/components/providers/*"],
             "@libs/*": ["./src/libs/*"],
